+++ src/Nirum/Targets/Python.hs
@@ -2,6 +2,7 @@
 {-# LANGUAGE ExtendedDefaultRules #-}
 {-# LANGUAGE OverloadedLists #-}
 {-# LANGUAGE QuasiQuotes #-}
+{-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TupleSections #-}
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeSynonymInstances #-}
@@ -12,7 +13,6 @@ module Nirum.Targets.Python
     , compileModule
     , compileTypeDeclaration
     , parseModulePath
-    , toNamePair
     ) where
 
 import Control.Monad (forM)
@@ -38,9 +38,10 @@ import Text.InterpolatedString.Perl6 (q, qq)
 import qualified Nirum.Constructs.Annotation as A
 import Nirum.Constructs.Annotation.Internal hiding (Text, annotations, name)
 import qualified Nirum.Constructs.Annotation.Internal as AI
+import Nirum.Constructs.Declaration hiding (annotations, name)
+import qualified Nirum.Constructs.Declaration
 import qualified Nirum.Constructs.DeclarationSet as DS
 import qualified Nirum.Constructs.Identifier as I
-import Nirum.Constructs.Declaration (Documented (docsBlock))
 import Nirum.Constructs.Module hiding (imports)
 import Nirum.Constructs.ModulePath ( ModulePath
                                    , fromIdentifiers
@@ -127,9 +128,6 @@ toEnumMemberName name'
     attributeName :: T.Text
     attributeName = toAttributeName' name'
 
-toNamePair :: Name -> T.Text
-toNamePair (Name f b) = [qq|('{toAttributeName f}', '{I.toSnakeCaseText b}')|]
-
 toIndentedCodes :: (a -> T.Text) -> [a] -> T.Text -> T.Text
 toIndentedCodes f traversable concatenator =
     T.intercalate concatenator $ map f traversable
@@ -198,6 +196,33 @@ compileDocstringWithFields indentSpace decl fields' =
                                      ]
         | f@(Field n _ _) <- toList fields'
         ]
+
+compileDocstringWithParameters :: forall a b . (Documented a, Declaration b)
+                               => Code
+                               -> a
+                               -> [(b, Code)]
+                               -> Code
+compileDocstringWithParameters indentSpace decl paramTypePairs =
+    compileDocstring'
+        indentSpace
+        decl
+        [ case compileDocs p of
+              Nothing ->
+                  [qq|:param {paramName p}: (:class:`{simplifyTypeExpr tx}`)
+|]
+              Just docs' ->
+                  [qq|:param {paramName p}: (:class:`{simplifyTypeExpr tx}`)
+{indentN (9 + (T.length $ paramName p)) docs'}
+|]
+        | (p, tx) <- paramTypePairs
+        ]
+  where
+    indentN :: Int -> Code -> Code
+    indentN n = indent (T.replicate n " ")
+    paramName :: b -> Code
+    paramName = toAttributeName' . Nirum.Constructs.Declaration.name
+    simplifyTypeExpr :: Code -> Code
+    simplifyTypeExpr = T.takeWhileEnd (/= '.')
 
 compileDocsComment :: Documented a => Code -> a -> Code
 compileDocsComment indentSpace d =
@@ -1031,74 +1056,105 @@ class #{className}(#{T.intercalate "," $ compileExtendClasses annotations}):
             | (A.Annotation annotationName _) <- A.toList annotations'
             ]
 
-compileTypeDeclaration
-    src@Source { sourcePackage = Package { metadata = metadata' } }
-    d@ServiceDeclaration { serviceName = name'
-                         , service = Service methods
-                         } = do
-    let methods' = toList methods
-    methodMetadata <- mapM compileMethodMetadata methods'
-    let methodMetadata' = commaNl methodMetadata
-    dummyMethods <- mapM compileMethod methods'
-    clientMethods <- mapM compileClientMethod methods'
-    methodErrorTypes <- mapM compileErrorType methods'
-    let dummyMethods' = T.intercalate "\n\n" dummyMethods
-        clientMethods' = T.intercalate "\n\n" clientMethods
-        methodErrorTypes' =
-            T.intercalate "," $ catMaybes methodErrorTypes
-    param <- parameterCompiler
-    ret <- returnCompiler
-    insertStandardImport "json"
-    insertThirdPartyImports [ ("nirum.deserialize", ["deserialize_meta"])
-                            ]
+compileTypeDeclaration src d@ServiceDeclaration { serviceName = name'
+                                                , service = Service methods
+                                                } = do
+    clientMethods <- mapM compileClientMethod $ toList methods
+    let clientMethods' = T.intercalate "\n\n" clientMethods
     insertThirdPartyImportsA
         [ ("nirum.constructs", [("name_dict_type", "NameDict")])
         , ("nirum.datastructures", [(nirumMapName, "Map")])
         , ("nirum.service", [("service_type", "Service")])
         , ("nirum.transport", [("transport_type", "Transport")])
         ]
-    return [qq|
-class $className(service_type):
-{compileDocstring "    " d}
+    pyVer <- getPythonVersion
+    methods' <- sequence $ (`map` toList methods) $ \ method' -> do
+        rTypeExpr <- compileTypeExpression' src $ returnType method'
+        errTypeExpr <- compileTypeExpression' src $ errorType method'
+        params' <- sequence $ (`map` toList (parameters method')) $
+            \ param'@(Parameter _ pType _) -> do
+                typeExpr <- compileTypeExpression' src $ Just pType
+                return (param', typeExpr)
+        return (method', rTypeExpr, errTypeExpr, params')
+    return $ toStrict $ renderMarkup [compileText|
+class #{className}(service_type):
+#{compileDocstring "    " d}
     __nirum_type__ = 'service'
-    __nirum_schema_version__ = \'{SV.toText $ version metadata'}\'
-    __nirum_service_methods__ = \{
-        {methodMetadata'}
-    \}
+    __nirum_service_methods__ = {
+%{ forall (Method mName _ _ _ _, rTypeExpr, _, params') <- methods' }
+        '#{toAttributeName' mName}': {
+            '_v': 2,
+            '_return': lambda: #{rTypeExpr},
+            '_names': name_dict_type([
+%{ forall (Parameter (Name pF pB) _ _, _) <- params' }
+                ('#{toAttributeName pF}', '#{I.toSnakeCaseText pB}'),
+%{ endforall }
+            ]),
+%{ forall (Parameter pName _ _, pTypeExpr) <- params' }
+            '#{toAttributeName' pName}': lambda: #{pTypeExpr},
+%{ endforall }
+        },
+%{ endforall }
+    }
     __nirum_method_names__ = name_dict_type([
-        $methodNameMap
+%{ forall (Method (Name mF mB) _ _ _ _, _, _, _) <- methods' }
+        ('#{toAttributeName mF}', '#{I.toSnakeCaseText mB}'),
+%{ endforall }
     ])
-    __nirum_method_annotations__ = $methodAnnotations'
+    __nirum_method_annotations__ = #{methodAnnotations'}
 
     @staticmethod
     def __nirum_method_error_types__(k, d=None):
-        return dict([
-            $methodErrorTypes'
-        ]).get(k, d)
+%{ forall (Method mName _ _ _ _, _, errTypeExpr, _) <- methods' }
+        if k == '#{toAttributeName' mName}':
+            return #{errTypeExpr}
+%{ endforall }
+        return d
 
-    {dummyMethods'}
+%{ forall (m@(Method mName _ _ _ _), rTypeExpr, _, params') <- methods' }
+    def #{toAttributeName' mName}(
+        self,
+%{ case pyVer }
+%{ of Python3 }
+%{ forall (Parameter pName _ _, pTypeExpr) <- params' }
+        #{toAttributeName' pName}: '#{pTypeExpr}',
+%{ endforall }
+    ) -> '#{rTypeExpr}':
+%{ of Python2 }
+%{ forall (Parameter pName _ _, _) <- params' }
+        #{toAttributeName' pName},
+%{ endforall }
+    ):
+%{ endcase }
+#{compileDocstringWithParameters "        " m params'}
+        raise NotImplementedError(
+            '#{className} has to implement #{toAttributeName' mName}()'
+        )
+%{ endforall }
 
 
-# FIXME client MUST be generated & saved on diffrent module
-#       where service isn't included.
-class {className}_Client($className):
+class #{className}_Client(#{className}):
     """The client object of :class:`{className}`."""
 
-    def __init__(self,
-                 { param "transport" "transport_type" }){ ret "None" }:
+%{ case pyVer }
+%{ of Python3 }
+    def __init__(self, transport: transport_type) -> None:
+%{ of Python2 }
+    def __init__(self, transport):
+%{ endcase }
         if not isinstance(transport, transport_type):
             raise TypeError(
-                'expected an instance of \{0.__module__\}.\{0.__name__\}, not '
-                '\{1!r\}'.format(transport_type, transport)
+                'expected an instance of {0.__module__}.{0.__name__}, not '
+                '{1!r}'.format(transport_type, transport)
             )
         self.__nirum_transport__ = transport  # type: transport_type
 
-    {clientMethods'}
+    #{clientMethods'}
 
-{className}.Client = {className}_Client
-{className}.Client.__name__ = 'Client'
-if hasattr({className}.Client, '__qualname__'):
-    {className}.Client.__qualname__ = '{className}.Client'
+#{className}.Client = #{className}_Client
+#{className}.Client.__name__ = 'Client'
+if hasattr(#{className}.Client, '__qualname__'):
+    #{className}.Client.__qualname__ = '#{className}.Client'
 |]
   where
     nirumMapName :: T.Text
@@ -1107,68 +1163,11 @@ if hasattr({className}.Client, '__qualname__'):
     className = toClassName' name'
     commaNl :: [T.Text] -> T.Text
     commaNl = T.intercalate ",\n"
-    compileErrorType :: Method -> CodeGen (Maybe Code)
-    compileErrorType (Method mn _ _ me _) =
-        case me of
-            Just errorTypeExpression -> do
-                et <- compileTypeExpression' src (Just errorTypeExpression)
-                return $ Just [qq|('{toAttributeName' mn}', $et)|]
-            Nothing -> return Nothing
-    compileMethod :: Method -> CodeGen Code
-    compileMethod m@(Method mName params rtype _etype _anno) = do
-        let mName' = toAttributeName' mName
-        params' <- mapM compileMethodParameter $ toList params
-        let paramDocs = [ T.concat [ ":param "
-                                   , toAttributeName' pName
-                                   , maybe "" (T.append ": ") $ compileDocs p
-                                   -- TODO: types
-                                   ]
-                        | p@(Parameter pName _ _) <- toList params
-                        ]
-        rtypeExpr <- compileTypeExpression' src rtype
-        ret <- returnCompiler
-        return [qq|
-    def {mName'}(self, {commaNl params'}){ ret rtypeExpr }:
-{compileDocstring' "        " m paramDocs}
-        raise NotImplementedError('$className has to implement {mName'}()')
-|]
     compileMethodParameter :: Parameter -> CodeGen Code
     compileMethodParameter (Parameter pName pType _) = do
         pTypeExpr <- compileTypeExpression' src (Just pType)
         arg <- parameterCompiler
         return [qq|{arg (toAttributeName' pName) pTypeExpr}|]
-    compileMethodMetadata :: Method -> CodeGen Code
-    compileMethodMetadata Method { methodName = mName
-                                 , parameters = params
-                                 , returnType = rtype
-                                 } = do
-        let params' = toList params :: [Parameter]
-        rtypeExpr <- compileTypeExpression' src rtype
-        paramMetadata <- mapM compileParameterMetadata params'
-        let paramMetadata' = commaNl paramMetadata
-        insertThirdPartyImportsA
-            [("nirum.constructs", [("name_dict_type", "NameDict")])]
-        return [qq|'{toAttributeName' mName}': \{
-            '_v': 2,
-            '_return': lambda: $rtypeExpr,
-            '_names': name_dict_type([{paramNameMap params'}]),
-            {paramMetadata'}
-        \}|]
-    methodList :: [Method]
-    methodList = toList methods
-    compileParameterMetadata :: Parameter -> CodeGen Code
-    compileParameterMetadata (Parameter pName pType _) = do
-        let pName' = toAttributeName' pName
-        pTypeExpr <- compileTypeExpression' src (Just pType)
-        return [qq|'{pName'}': lambda: $pTypeExpr|]
-    methodNameMap :: T.Text
-    methodNameMap = toIndentedCodes
-        toNamePair
-        [mName | Method { methodName = mName } <- methodList]
-        ",\n        "
-    paramNameMap :: [Parameter] -> T.Text
-    paramNameMap params = toIndentedCodes
-        toNamePair [pName | Parameter pName _ _ <- params] ",\n        "
     compileClientPayload :: Parameter -> CodeGen Code
     compileClientPayload (Parameter pName pt _) = do
         let pName' = toAttributeName' pName
@@ -1292,7 +1291,7 @@ if hasattr({className}.Client, '__qualname__'):
             ]
     methodAnnotations' :: T.Text
     methodAnnotations' = wrapMap $ commaNl $ map compileMethodAnnotation
-        methodList
+        (toList methods)
 
 compileTypeDeclaration _ Import {} =
     return ""  -- Nothing to compile